[SEMVER-MAJOR] Stop gzipping CLDR files
Rework the module to use plain JSON format for CLDR files and thus get
rid of dependency on zlib's gunzipAsync that's not available on Node
v0.10 on Windows.

Breaking change: when scanning for custom CLDR files, strong-globalize
we no longer recogizes `.gz` files. If you are relying on this feature,
then ungzip your CLDR files and rename them to `.json`.