remove references to gzip'd CLDR data
This involves a bunch of somewhat unrelated changes that for various
reasons cannot be made separately.

Firstly, don't prune alternate CLDR formats during install. The gains
are negligible considering the cost of bandwidth and disk writing has
already been paid.

Additionally, the postinstall prevents a vanilla 'npm install && npm
test && npm publish' workflow from publishing a valid package because it
prunes the CLDR data based on the version of node used to publish the
module.

This fixes #51

Secondly, we need to fix the problem of node 0.10 seeing the gz version
of the file because if it tries to load it, it will bail because it
does't have a synchronous gunzip API. To fix that, we have to prevent
strong-globalize from even looking at the gz version of the file.

Thirdly, we need to update the various bits of tests that were in place
to support this gz/non-gz behaviour split.

Now we have a bunch of dead code and an unuzed gz file, so remove them.

And that is how we end up with this larger-than-it-should-be commit
instead of a bunch of discrete changes.