test: fix empty subtest in TAP output
When run in the master process scenario, the subtest had 0 assertions in
it. The resulting TAP output confuses tap4j causing all sorts of fun.

Only load tap in the child process so that the only TAP that is seen by
the runner is coming from the child process and not double-interpretted
by tap being loaded into the parent process.

This entire test is way more complicated than it should be. I'm not sure
what it is supposed to be testing, but it sure doesn't fit with how
node-tap was meant to work.